LayoutTests/js/sparse-array-expected.txt

This tests some sparse array operations.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS array[0] is undefined
PASS array[49999] is undefined
PASS array[50000] is 100
PASS array[50001] is undefined
PASS array[0] is NaN
PASS array[49999] is undefined
PASS array[50000] is 100
PASS array[50001] is undefined
PASS array[0] is NaN
PASS array[49999] is undefined
PASS array[50000] is 100
PASS array[50001] is undefined

PASS successfullyParsed is true

TEST COMPLETE
